Move the dialog to chapter package.
