+++ app/src/main/java/org/gdg/frisbee/android/chapter/MainActivity.java
@@ -55,7 +55,6 @@ import org.gdg.frisbee.android.eventseries.GdgEventListFragment;
 import org.gdg.frisbee.android.onboarding.FirstStartActivity;
 import org.gdg.frisbee.android.utils.PrefUtils;
 import org.gdg.frisbee.android.utils.Utils;
-import org.gdg.frisbee.android.widget.ChapterSelectDialog;
 import org.joda.time.DateTime;
 
 import java.util.ArrayList;
